// src/shared/enums/core-repository.enum.ts
export enum CoreRepositoryEnum {
  ATTENDANCE_REPOSITORY = 'ATTENDANCE_REPOSITORY',
  HORARIO_REPOSITORY = 'HORARIO_REPOSITORY',
  PERMISSION_REPOSITORY = 'PERMISSION_REPOSITORY',
  PERMISSION_TYPE_REPOSITORY = 'PERMISSION_TYPE_REPOSITORY',
  VACATION_REPOSITORY = 'VACATION_REPOSITORY',
  FORM_REPOSITORY = 'FORM_REPOSITORY',
  PERMISSION_STATE_REPOSITORY = 'PERMISSION_STATE_REPOSITORY',
  VACATION_DETAIL_REPOSITORY = 'VACATION_DETAIL_REPOSITORY',
  SIGNER_REPOSITORY = 'SIGNER_REPOSITORY',
  SITE_REPOSITORY = 'SITE_REPOSITORY',
  EMPLOYEE_REPOSITORY = 'EMPLOYEE_REPOSITORY',
  ASISTENCIA_REPOSITORY = 'ASISTENCIA_REPOSITORY',
} 